requirements.txt
discord.py==2.3.2
aiosqlite==0.19.0
APScheduler==3.10.4
python-dotenv==1.0.0
Pillow==12.2.0
aiohttp==3.9.5
